improve: 删除活动二次确认

+++ ylcs-app/src/commonMain/kotlin/love/yinlin/ui/screen/world/ScreenActivityDetails.kt
@@ -242,7 +242,9 @@ class ScreenActivityDetails(model: AppModel, private val args: Args) : SubScreen
 				navigate(ScreenModifyActivity.Args(args.aid))
 			}
 			ActionSuspend(Icons.Outlined.Delete) {
-				deleteActivity()
+				if (slot.confirm.openSuspend(content = "删除活动")) {
+					deleteActivity()
+				}
 			}
 		}
 	}
